backend/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-rest==0.8.7
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2